requirements.txt
# Python dependencies for Vercel serverless functions
google-generativeai
python-dotenv
firebase-admin
Pillow
requests
